connect Product page to Goods page
