+++ client/src/App.js
@@ -2,8 +2,8 @@ import {Route, Switch} from "react-router-dom";
 import NotFound from "./components/pages/NotFound";
 import Home from "./components/pages/Home";
 import Goods from "./components/pages/Goods";
-import Product from "./components/assets/Product/Product";
 import {ProductProvider} from "./components/assets/Product/ProductContext";
+import ProductPage from "./components/pages/ProductPage";
 
 
 function App() {
@@ -12,7 +12,7 @@ function App() {
             <Switch>
                 <Route exact path="/" component={Home}/>
                 <Route exact path="/goods/:product_type" component={Goods}/>
-                <Route path="/goods/:product_type/:id" component={Product}/>
+                <Route path="/goods/:product_type/:id" component={ProductPage}/>
                 <Route component={NotFound}/>
             </Switch>
         </ProductProvider>
